Don't mutate datetime (#167)
* no mutation

* refactor modules + add test

* Add moment-based timestamps

* Export getCurrentDate()
