+++ packages/mds-utils/tests/utils.spec.ts
@@ -17,16 +17,7 @@
 import test from 'unit.js'
 import assert from 'assert'
 import { VEHICLE_EVENTS, VehicleEvent } from '@mds-core/mds-types'
-import {
-  routeDistance,
-  filterEmptyHelper,
-  parseOperator,
-  parseCount,
-  parseUnit,
-  parseIsRelative,
-  isStateTransitionValid,
-  normalizeToArray
-} from '../utils'
+import { routeDistance, filterEmptyHelper, isStateTransitionValid, normalizeToArray } from '../utils'
 import { expectedTransitions } from './state-transition-expected'
 
 const Boston = { lat: 42.360081, lng: -71.058884 }
@@ -80,62 +71,6 @@ describe('Tests Utilities', () => {
     //   Sinon.restore()
     //   log.warn = oldLogWarn
     // })
-  })
-
-  describe('Date/time API utils', () => {
-    describe('Parses operators', () => {
-      it('Gets the operator', () => {
-        assert.strictEqual(parseOperator('+5d'), '+')
-        assert.strictEqual(parseOperator('-49d'), '-')
-        assert.strictEqual(parseOperator('today'), '+')
-        assert.strictEqual(parseOperator('yesterday'), '+')
-        assert.strictEqual(parseOperator('now'), '+')
-      })
-
-      it('Rejects malformed strings', () => {
-        assert.throws(() => parseOperator('bad-offset'))
-      })
-    })
-
-    describe('Parses counts', () => {
-      it('Parses counts', () => {
-        assert.strictEqual(parseCount('+5d'), 5)
-        assert.strictEqual(parseCount('-49d'), 49)
-        assert.strictEqual(parseCount('today'), 0)
-        assert.strictEqual(parseCount('yesterday'), 1)
-        assert.strictEqual(parseCount('now'), 0)
-      })
-
-      it('Rejects malformed strings', () => {
-        assert.throws(() => parseCount('bad-offset'))
-      })
-    })
-
-    describe('Parses units', () => {
-      it('Parses units', () => {
-        assert.strictEqual(parseUnit('+5d'), 'days')
-        assert.strictEqual(parseUnit('-49d'), 'days')
-        assert.strictEqual(parseUnit('-49h'), 'hours')
-        assert.strictEqual(parseUnit('today'), 'days')
-        assert.strictEqual(parseUnit('yesterday'), 'days')
-        assert.strictEqual(parseUnit('now'), 'days')
-      })
-
-      it('Rejects malformed strings', () => {
-        assert.throws(() => parseUnit('bad-offset'))
-      })
-    })
-
-    describe('Parses is relative', () => {
-      it('Parses is relative', () => {
-        assert.strictEqual(parseIsRelative('+5d'), true)
-        assert.strictEqual(parseIsRelative('-49d'), true)
-        assert.strictEqual(parseIsRelative('-49h'), true)
-        assert.strictEqual(parseIsRelative('today'), false)
-        assert.strictEqual(parseIsRelative('yesterday'), false)
-        assert.strictEqual(parseIsRelative('now'), false)
-      })
-    })
   })
 
   describe('Normalize to array', () => {
